+++ libft/ft_split.c
@@ -6,7 +6,7 @@
 /*   By: trolland <trolland@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/30 12:40:00 by trolland          #+#    #+#             */
-/*   Updated: 2024/01/12 17:26:36 by trolland         ###   ########.fr       */
+/*   Updated: 2024/02/24 10:26:41 by trolland         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ static char	*ft_create_sub_tab(char const *s, char c, int *pos)
 		(*pos)++;
 	while (ft_is_charset(s[*pos + len], c) && s[*pos + len])
 		len++;
-	tab = malloc(sizeof(char) * len + 1);
+	tab = malloc(sizeof(char) * (unsigned long)(len + 1));
 	if (!tab)
 		return (NULL);
 	while (k < len)
@@ -79,7 +79,7 @@ char	**ft_split(char const *s, char c)
 	i = 0;
 	j = 0;
 	word_count = (ft_count_words_one_char(s, c));
-	split = malloc(sizeof(char *) * (word_count + 1));
+	split = malloc(sizeof(char *) * (unsigned long)(word_count + 1));
 	if (!split)
 		return (NULL);
 	while (i < word_count)
