+++ libft/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: trolland <trolland@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/31 15:16:09 by trolland          #+#    #+#             */
-/*   Updated: 2023/09/26 12:01:11 by trolland         ###   ########.fr       */
+/*   Updated: 2024/02/24 10:28:25 by trolland         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ char	*ft_strrchr(const char *s, int c)
 {
 	int	i;
 
-	i = ft_strlen(s);
+	i = (int)ft_strlen(s);
 	while (i >= 0 && s != NULL)
 	{
 		if (s[i] == (char)c)
