+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: trolland <trolland@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/31 16:52:28 by trolland          #+#    #+#             */
-/*   Updated: 2023/10/05 13:04:15 by trolland         ###   ########.fr       */
+/*   Updated: 2024/02/24 10:28:52 by trolland         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,12 +36,12 @@ char	*ft_strtrim(char const *s1, char const *set)
 	if (!s1)
 		return (NULL);
 	start = 0;
-	end = ft_strlen(s1);
+	end = (int)ft_strlen(s1);
 	while (s1[start] && ft_is_charset(s1[start], set))
 		start++;
 	while (s1[end - 1] && ft_is_charset(s1[end - 1], set) && end > start)
 		end--;
-	trim = malloc(sizeof(char) * (end - start + 1));
+	trim = malloc(sizeof(char) * (unsigned long)(end - start + 1));
 	if (!trim)
 		return (NULL);
 	i = 0;
